+++ tests/Unit/data/migrations_using_soft_deletes/2020_01_30_000000_create_users_table_with_soft_deletes.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Database\Migrations;
+namespace Tests\Unit\MigrationsUsingSoftDeletes;
 
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
@@ -17,16 +17,8 @@ class CreateUsersTable extends Migration
     {
         Schema::create('users', static function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->string('name');
+            $table->string('name')->nullable();
             $table->string('email')->unique();
-            $table->timestamp('email_verified_at')->nullable();
-            $table->string('password');
-            $table->json('meta');
-            $table->json('options');
-            $table->json('properties');
-            $table->boolean('blocked');
-            $table->unknownColumnType('unknown_column');
-            $table->rememberToken();
             $table->timestamps();
             $table->softDeletes();
         });