+++ tests/Unit/data/migrations_using_soft_deletes_tz/2020_01_31_000000_alter_users_table_to_use_soft_deletes_tz.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Database\Migrations;
+namespace Tests\Unit\MigrationsUsingSoftDeletesTz;
 
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
@@ -17,18 +17,10 @@ class CreateUsersTable extends Migration
     {
         Schema::create('users', static function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->string('name');
+            $table->string('name')->nullable();
             $table->string('email')->unique();
-            $table->timestamp('email_verified_at')->nullable();
-            $table->string('password');
-            $table->json('meta');
-            $table->json('options');
-            $table->json('properties');
-            $table->boolean('blocked');
-            $table->unknownColumnType('unknown_column');
-            $table->rememberToken();
             $table->timestamps();
-            $table->softDeletes();
+            $table->softDeletesTz();
         });
     }
 }