+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+* Fix: the `deleted_at` column created with `SoftDeletes` is now treated as a `Carbon` instance
+
 ## [2.0.1] - 2022-02-09
 
 ### Added